+++ acmecse/runtime/configurations/CNTResourceConfiguration.py
@@ -26,6 +26,7 @@ class CNTResourceConfiguration(ModuleConfiguration):
 		config.resource_cnt_mni = parser.getint('resource.cnt', 'mni', fallback=10)
 		config.resource_cnt_mbs = parser.getint('resource.cnt', 'mbs', fallback=10000)
 		config.resource_cnt_mia = parser.getint('resource.cnt', 'mia', fallback=60*60*24*365*5) # 5 years, in seconds
+		config.resource_cnt_mbis = parser.getint('resource.cnt', 'mbis', fallback=-1)
 
 
 	def validateConfiguration(self, config:Configuration, initial:Optional[bool]=False) -> None:
@@ -35,3 +36,6 @@ class CNTResourceConfiguration(ModuleConfiguration):
 			raise ConfigurationError(r'[i]\[resource.cnt]:mbs[/i] must be > 0')
 		if config.resource_cnt_mia <= 0:
 			raise ConfigurationError(r'[i]\[resource.cnt]:mia[/i] must be > 0')
+		if config.resource_cnt_mbis > config.resource_cnt_mbs:
+			raise ConfigurationError(r'[i]\[resource.cnt]:mbis[/i] cannot be greater than [i]\[resource.cnt]:mbs[/i]')
+